% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extents.R
\name{extents}
\alias{extents}
\title{All extents}
\usage{
extents(x)
}
\arguments{
\item{x}{Object with paths}
}
\value{
a dataframe of object and extent values (xmin, xmax, ymin, ymax)
}
\description{
(This function probably belongs in spex). Find the extent of all paths within an object.
}
\details{
The `path_` identifier is included, but won't be of use without an
existing `PATH` object. The path order is implicit as per the gibble
geometry map.
}
\examples{
data("minimal_mesh", package = "silicate")
extents(minimal_mesh)
}
