% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfft-package.R
\docType{package}
\name{pfft-package}
\alias{pfft}
\alias{pfft-package}
\title{pfft: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
